ai-core/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
torch==2.1.1
transformers==4.35.2
faiss-cpu==1.7.4
numpy==1.26.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
cryptography==41.0.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pymongo==4.6.1
redis==5.0.1
python-dotenv==1.0.0
